+++ client/src/store/entities/ammo.ts
@@ -16,18 +16,9 @@ class AmmoSupplier implements EntitySupplier {
 	}
 }
 
-// Refer to gun color for the order
-// Inner order: frame, outer, inner
-const COLOR_SCHEME = [
-	["#332300", "#b37a00", "#f2a500"],
-	["#4c0000", "#b30000", "#f20000"],
-	["#002c6d", "#0048b3", "#0061f2"],
-	["#013d00", "#026f00", "#039600"],
-	["#111111", "#111111", "#006400"],
-];
-
 export default class Ammo extends Entity {
 	static readonly TYPE = "ammo";
+	static colorScheme: string[][] = [];
 	type = Ammo.TYPE;
 	amount!: number;
 	color!: GunColor;
@@ -40,6 +31,7 @@ export default class Ammo extends Entity {
 
 	static {
 		ENTITY_SUPPLIERS.set(Ammo.TYPE, new AmmoSupplier());
+		fetch("data/colors/ammos.json").then(res => res.json()).then(x => this.colorScheme = x);
 	}
 
 	copy(minEntity: MinEntity & AdditionalEntity) {
@@ -54,12 +46,12 @@ export default class Ammo extends Entity {
 		ctx.rotate(-this.direction.angle());
 		ctx.scale(scale, scale);
 		const length = this.hitbox.comparable * Math.sin(CommonAngle.PI_TWO);
-		ctx.strokeStyle = COLOR_SCHEME[this.color][0];
+		ctx.strokeStyle = `#${Ammo.colorScheme[this.color][0]}`;
 		ctx.lineWidth = 0.2;
-		ctx.fillStyle = COLOR_SCHEME[this.color][1];
+		ctx.fillStyle = `#${Ammo.colorScheme[this.color][1]}`;
 		ctx.fillRect(-length / 2, -length / 2, length, length);
 		ctx.strokeRect(-length / 2, -length / 2, length, length);
-		ctx.fillStyle = COLOR_SCHEME[this.color][2];
+		ctx.fillStyle = `#${Ammo.colorScheme[this.color][2]}`;
 		ctx.fillRect(-length / 8, -length / 4, length / 3, length / 3);
 		ctx.resetTransform();
 	}
