+++ client/src/store/weapons/index.ts
@@ -1,4 +1,3 @@
-import { OPENSURVIV_DATA } from "../../constants";
 import { MeleeData, GunData } from "../../types/data";
 import { MinWeapon } from "../../types/minimized";
 import { WeaponSupplier } from "../../types/supplier";
@@ -41,13 +40,13 @@ class GunSupplier implements WeaponSupplier {
 }
 
 (async() => {
-	for (const file of await fetch(`${OPENSURVIV_DATA}/data/weapons/melee/.list.json`).then(res => res.json())) {
-		const data = <MeleeData> await fetch(`${OPENSURVIV_DATA}/data/weapons/melee/${file}.json`).then(res => res.json());
+	for (const file of await fetch(`data/weapons/melee/.list.json`).then(res => res.json()).catch(err => console.error(err))) {
+		const data = <MeleeData> await fetch(`data/weapons/melee/${file}.json`).then(res => res.json());
 		WEAPON_SUPPLIERS.set(file, new MeleeSupplier(file, data));
 	}
 	
-	for (const file of await fetch(`${OPENSURVIV_DATA}/data/weapons/guns/.list.json`).then(res => res.json())) {
-		const data = <GunData> await fetch(`${OPENSURVIV_DATA}/data/weapons/guns/${file}.json`).then(res => res.json())
+	for (const file of await fetch(`data/weapons/guns/.list.json`).then(res => res.json())) {
+		const data = <GunData> await fetch(`data/weapons/guns/${file}.json`).then(res => res.json())
 		WEAPON_SUPPLIERS.set(file, new GunSupplier(file, data));
 	}
 })();
