+++ client/src/textures.ts
@@ -1,9 +1,8 @@
-import { OPENSURVIV_DATA } from "./constants";
 import { TracerColor, TracerColorData } from "./types/data";
 
 const weapons = new Map<string, HTMLImageElement & { loaded: boolean }>();
 (async() => {
-	for (const id of await fetch(`${OPENSURVIV_DATA}/data/weapons/guns/.list.json`).then(res => res.json())) {
+	for (const id of await fetch(`data/weapons/guns/.list.json`).then(res => res.json())) {
 		const img: HTMLImageElement & { loaded: boolean } = Object.assign(new Image(), { loaded: false });
 		img.onload = () => img.loaded = true;
 		img.src = `assets/images/game/loots/weapons/${id}.png`;
@@ -18,7 +17,7 @@ export function getWeaponImage(id: string) {
 
 const tracerColors = new Map<string, TracerColor>();
 (async() => {
-	const data = <TracerColorData> await fetch(`${OPENSURVIV_DATA}/data/colors/tracers.json`).then(res => res.json());
+	const data = <TracerColorData> await fetch(`data/colors/tracers.json`).then(res => res.json());
 	for (const id of Object.keys(data)) {
 		tracerColors.set(id, data[id]);
 	}
