+++ data/amount/ammos.json
@@ -1,0 +1,6 @@
+[
+	[120, 15, 90, 90, 49, 10, 2, 90, 10, 0, 0, 1, 0, 0],
+	[240, 30, 180, 180, 98, 20, 4, 180, 20, 0, 0, 1, 0, 0],
+	[330, 60, 240, 240, 147, 40, 6, 240, 30, 0, 0, 1, 0, 0],
+	[420, 90, 300, 300, 196, 80, 8, 300, 40, 0, 0, 1, 0, 0]
+]
