+++ data/colors/ammos.json
@@ -1,0 +1,14 @@
+[
+	["332300", "b37a00", "f2a500"],
+	["4c0000", "b30000", "f20000"],
+	["002c6d", "0048b3", "0061f2"],
+	["013d00", "026f00", "039600"],
+	["101010", "222222", "2b2b2b"],
+	["121300", "272b00", "323800"],
+	["2e1600", "943400", "c74400"],
+	["1e0036", "5600b1", "7100f0"],
+	["022c22", "089b78", "0bd2a2"],
+	["170d06", "512c15", "6e3c1c"],
+	["333333", "dda9b2", "fbeff1"],
+	["000000", "000000", "000000"]
+]
