+++ data/weapons/guns/.list.json
