+++ data/weapons/guns/.template.json
