+++ data/weapons/guns/ak47.json
