+++ data/weapons/guns/awm.json
