+++ data/weapons/guns/mosin_nagant.json
