+++ data/weapons/guns/mp5.json
