+++ data/weapons/guns/sv98.json
