+++ data/weapons/melee/.template.json
